refactor($test) : 테스트 파일 패키지 변경
LottoTest,WinningNumbersTest -> domain
ParserTest,ValidatorTest -> util
패키지로 변경
