+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,16 +1,11 @@
-package lotto;
+package lotto.domain;
 
 import java.util.Arrays;
-import java.util.stream.Collectors;
 import lotto.constant.Rank;
-import lotto.domain.Lotto;
-import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
