+++ src/test/java/lotto/util/ParserTest.java
@@ -1,6 +1,5 @@
-package lotto;
+package lotto.util;
 
-import lotto.util.Parser;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
